+++ start/backendServices/src/main/java/io/openliberty/guides/inventory/InventoryResource.java
@@ -24,7 +24,6 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import io.openliberty.guides.inventory.model.InventoryList;
 import javax.annotation.security.RolesAllowed;
-import javax.annotation.security.DeclareRoles;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.Context;
 
@@ -42,6 +41,7 @@ public class InventoryResource {
       @Context HttpHeaders httpHeaders) {
     String authHeader = httpHeaders.getRequestHeaders()
                                    .getFirst(HttpHeaders.AUTHORIZATION);
+    // Get properties
     Properties props = manager.get(hostname, authHeader);
     if (props == null) {
       return Response.status(Response.Status.NOT_FOUND)
@@ -49,6 +49,9 @@ public class InventoryResource {
                          "ERROR: Unknown hostname or the resource may not be running on the host machine")
                      .build();
     }
+
+    // Add to inventory
+    manager.add(hostname, props);
     return Response.ok(props).build();
   }
